refactor(ReservationTime): 어드민과 유저에 맞게 기능 분리
