--- src/main/java/roomescape/reservationtime/ReservationTimeController.java
@@ -1,59 +1,0 @@
-package roomescape.reservationtime;
-
-import java.time.LocalDate;
-import java.util.List;
-
-import jakarta.validation.Valid;
-import org.springframework.format.annotation.DateTimeFormat;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/times")
-public class ReservationTimeController {
-    private final ReservationTimeService reservationTimeService;
-
-    public ReservationTimeController(ReservationTimeService reservationTimeService) {
-        this.reservationTimeService = reservationTimeService;
-    }
-
-    @GetMapping
-    public ResponseEntity<List<ReservationTimeResponse>> getReservationTimes() {
-        List<ReservationTime> reservationTimes = reservationTimeService.findReservationTimes();
-        List<ReservationTimeResponse> response = reservationTimes.stream()
-                .map(ReservationTimeResponse::from)
-                .toList();
-
-        return ResponseEntity.ok(response);
-    }
-
-    @PostMapping
-    public ResponseEntity<ReservationTimeResponse> createReservationTime(
-            @Valid @RequestBody ReservationTimeRequest reservationTimeRequest) {
-        ReservationTime reservationTime = reservationTimeService.createReservationTime(reservationTimeRequest.startAt());
-        ReservationTimeResponse response = ReservationTimeResponse.from(reservationTime);
-        return ResponseEntity.status(HttpStatus.CREATED).body(response);
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteReservationTime(@PathVariable long id) {
-        reservationTimeService.deleteReservationTime(id);
-        return ResponseEntity.noContent().build();
-    }
-
-    @PostMapping("/{themeId}")
-    public ResponseEntity<ScheduleResponse> getSchedules(@PathVariable long themeId,
-                                                         @RequestParam @DateTimeFormat(pattern = "yyyy-MM-dd") LocalDate date) {
-
-        ScheduleResponse response = reservationTimeService.getSchedules(date, themeId);
-        return ResponseEntity.ok().body(response);
-    }
-}
